Moving internal `mptrie.Proof` struct to pkg `state.Proof` with all relevant methods
